generation fix

+++ unit/strategy.py
@@ -3,5 +3,5 @@ from dataclasses import dataclass
 
 @dataclass
 class Strategy:
-    closer: int = 0
-    weaker: int = 0
+    closer: float = 0
+    weaker: float = 0
